requirements.txt
requests>=2.28.0
pandas>=1.5.0
tqdm>=4.64.0
psycopg2-binary>=2.9.5
sqlalchemy>=1.4.0
python-dotenv>=1.0.0 